% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xhash.R
\name{hash-ops}
\alias{[.xhash}
\alias{[<-.xhash}
\alias{hash-ops}
\alias{haskey}
\alias{haskey.xhash}
\title{Hash Operations for Single State}
\usage{
haskey(x, ...)

\method{[}{xhash}(x, state, ...)

\method{[}{xhash}(x, state, ...) <- value

\method{haskey}{xhash}(x, state, ...)
}
\arguments{
\item{x}{object}

\item{...}{additional arguments to determine the key}

\item{state}{state object}

\item{value}{value to assign}
}
\value{
\itemize{
  \item{\code{haskey} returns a logical}
  \item{\code{`[`} returns a reference to the object}
  \item{\code{`[<-`} returns a value}
}
}
\description{
Hash Operations for Single State
}

