% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_ai.R
\name{ttt_ai}
\alias{ttt_ai}
\title{Tic-Tac-Toe AI Player}
\usage{
ttt_ai(name = "ttt AI", value_func = NULL, policy_func = NULL)
}
\arguments{
\item{name}{player name}

\item{value_func}{\code{xhash} object that maps game state onto value}

\item{policy_func}{\code{xhash} object that maps game state onto action set}
}
\value{
\code{ttt_ai} object
}
\description{
Tic-Tac-Toe AI Player
}

