% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_human.R
\name{ttt_human}
\alias{ttt_human}
\title{Human Tic-Tac-Toe Player}
\usage{
ttt_human(name = "no name")
}
\arguments{
\item{name}{player name}
}
\value{
\code{ttt_human} object
}
\description{
Create an human tic-tac-toe player
}

