% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_randbot.R
\name{ttt_randbot}
\alias{ttt_randbot}
\title{Random AI Player of Tic-Tac-Toe Game}
\usage{
ttt_randbot(name = "random AI")
}
\arguments{
\item{name}{player name}
}
\value{
\code{ttt_randbot} object
}
\description{
Random AI Player of Tic-Tac-Toe Game
}

