% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_simulate.R
\name{ttt_simulate}
\alias{ttt_simulate}
\title{Simulate Tic-Tac-Toe Games between AIs}
\usage{
ttt_simulate(player1, player2 = player1, N = 1000L, verbose = TRUE)
}
\arguments{
\item{player1, player2}{AI players to simulate}

\item{N}{number simulation games}

\item{verbose}{if true, show progress report}
}
\value{
integer vector of simulation outcomes
}
\description{
Simulate Tic-Tac-Toe Games between AIs
}
\examples{
res <- ttt_simulate(ttt_ai(), ttt_ai())
prop.table(table(res))
}

