% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xhash.R
\name{vectorized-hash-ops}
\alias{vectorized-hash-ops}
\alias{haskeys}
\alias{setvalues}
\alias{getvalues}
\alias{getvalues.xhash}
\alias{setvalues.xhash}
\alias{haskeys.xhash}
\title{Vectorized Hash Operations}
\usage{
haskeys(x, ...)

setvalues(x, ...)

getvalues(x, ...)

\method{getvalues}{xhash}(x, states, ...)

\method{setvalues}{xhash}(x, states, values, ...)

\method{haskeys}{xhash}(x, states, ...)
}
\arguments{
\item{x}{object}

\item{...}{additional arugments to determine the keys}

\item{states}{state object}

\item{values}{values to assign}
}
\value{
\itemize{
  \item{\code{haskeys} returns a logical vector}
  \item{\code{setvalues} returns a reference to the object}
  \item{\code{getvalues} returns a list of values}
}
}
\description{
Vectorized Hash Operations
}
