% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_ai.R
\name{ttt_ai}
\alias{ttt_ai}
\title{Tic-Tac-Toe AI Player}
\usage{
ttt_ai(name = "ttt AI", level = 0L)
}
\arguments{
\item{name}{player name}

\item{level}{AI strength. must be Integer 0 (weekest) to 5 (strongest)}
}
\value{
\code{ttt_ai} object
}
\description{
Tic-Tac-Toe AI Player
}

