% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_ai.R
\name{ttt_ai}
\alias{ttt_ai}
\alias{ttt_randbot}
\title{Tic-Tac-Toe AI Player}
\usage{
ttt_ai(name = "ttt AI", level = 0L)

ttt_randbot(name = "random AI")
}
\arguments{
\item{name}{player name}

\item{level}{AI strength. must be Integer 0 (weekest) to 5 (strongest)}
}
\value{
\code{ttt_ai} object
}
\description{
Create an AI tic-tac-toe game player
}
\details{
\code{ttt_randbot} is an alias of \code{ttt_ai(level = 0)}.
}

