% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_game.R
\name{ttt_game}
\alias{ttt_game}
\title{Tic-Tac-Toe Game}
\usage{
ttt_game()
}
\value{
\code{ttt_game} object
}
\description{
Object that encapsulates a tic-tac-toe game.
}
\examples{
x <- ttt_game()
x$play(3)
x$play(5)
x$show_board()
}

