% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_human.R
\name{ttt_human}
\alias{ttt_human}
\title{Human Tic-Tac-Toe Player}
\usage{
ttt_human(name = "no name")
}
\arguments{
\item{name}{player name}
}
\value{
\code{ttt_human} object

\strong{Fields}
\describe{
  \item{\code{name}}{Player name}
}

\strong{Methods}
\describe{
  \item{\code{getmove(game, prompt = "choose move (e.g. A1) > ", ...)}}{Communicate with users to type in the next move.

    \emph{Input:}
    \itemize{
      \item{\code{game}: \code{\link{ttt_game}} object}
      \item{\code{prompt}: prompt message}
    }

    \emph{Output:} a character of a move
  }
}
}
\description{
Create an human tic-tac-toe player
}
\examples{
\dontrun{
p <- ttt_human()
p$getmove()
}
}
