% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_qlearn.R
\name{ttt_qlearn}
\alias{ttt_qlearn}
\title{Q-Learning to Train Tic-Tac-Toe AI}
\usage{
ttt_qlearn(player, N = 1000L, epsilon = 0.1, alpha = 0.8, gamma = 0.99,
  simulate = TRUE, sim_every = 250L, N_sim = 1000L, verbose = TRUE)
}
\arguments{
\item{player}{AI player to train}

\item{N}{number of episode, i.e. training games}

\item{epsilon}{fraction of random exploration}

\item{alpha}{learning rate}

\item{gamma}{discount factor}

\item{simulate}{if true, conduct simulation during training}

\item{sim_every}{conduct simulation after this many episode}

\item{N_sim}{number of simulation games}

\item{verbose}{if true, progress report is shown}
}
\value{
\code{data.frame} of simulation outcomes if any
}
\description{
Q-Learning to Train Tic-Tac-Toe AI
}
\examples{
p <- ttt_ai()
o <- ttt_qlearn(p, N = 200)
}

