% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttt_simulate.R
\name{ttt_simulate}
\alias{ttt_simulate}
\title{Simulate Tic-Tac-Toe Games between AIs}
\usage{
ttt_simulate(player1, player2 = player1, N = 1000L, verbose = TRUE,
  showboard = FALSE, pauseif = integer(0))
}
\arguments{
\item{player1, player2}{AI players to simulate}

\item{N}{number of simulation games}

\item{verbose}{if true, show progress report}

\item{showboard}{if true, game transition is displayed}

\item{pauseif}{pause the simulation when specified results occur.
This can be useful for explorative purposes.}
}
\value{
integer vector of simulation outcomes
}
\description{
Simulate Tic-Tac-Toe Games between AIs
}
\examples{
res <- ttt_simulate(ttt_ai(), ttt_ai())
prop.table(table(res))
}

